Replaced the fragments
