+++ app/src/main/java/com/appTec/RegistrateApp/view/fragments/device2/DeviceView2.java
@@ -1,10 +1,10 @@
-package com.appTec.RegistrateApp.view.fragments.device;
+package com.appTec.RegistrateApp.view.fragments.device2;
 
 import com.appTec.RegistrateApp.models.Device;
 
 import java.util.ArrayList;
 
-public interface DeviceView {
+public interface DeviceView2 {
     /**
      * Device View
      */
