--- app/src/main/java/com/appTec/RegistrateApp/view/fragments/permission/PermissionView.java
@@ -1,4 +1,0 @@
-package com.appTec.RegistrateApp.view.fragments.permission;
-
-public interface PermissionView {
-}
