+++ app/src/main/java/com/appTec/RegistrateApp/view/fragments/permission2/PermissionView2.java
@@ -1,0 +1,4 @@
+package com.appTec.RegistrateApp.view.fragments.permission2;
+
+public interface PermissionView2 {
+}
