+++ app/src/main/res/layout/fragment_permission.xml
@@ -1,31 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!-- Fragment to show notifications -->
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">z
+    android:layout_height="match_parent">
 
     <ListView
-        android:id="@+id/lvPermission"
+        android:id="@+id/permission_list_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        android:dividerHeight="30dp"
-        android:divider="@drawable/listview_divider">
-    </ListView>
+        android:visibility="visible"/>
 
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fabAddPermission"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:clickable="true"
-        app:srcCompat="@drawable/ic_add"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginBottom="10dp"
-        android:layout_marginRight="10dp" />
 
 </RelativeLayout>
